move formatDate and formatDateRange to DateFormatUtils
